requirements.txt
diffusers==0.34.0.dev0
numpy==2.3.1
Pillow==11.2.1
torch==2.7.1+cu118
